Introduce AST for validation expressions
